Implement the package according to the spec
Fixes https://github.com/kyrylo/netstring/issues/4
(A netstring carries size information. It is encoded as 4-byte uint32. Why?)

We were calculating the netstring length wrong. Packing also used to work not
according to the netstring spec. We used to assume that netstring length is
encoded as a 4-byte uint32, while the spec clearly says that it is encoded as
ASCII digits. Every bit in the length field means a digit, not a part of a
length number.

Big thanks to the caring people who pointed that out. In this, almost complete,
rewrite, we make sure that we calculate digits correctly.